SI-6054 Modern eta-exp examples that almost run
This wasn't so easy after all. The clever one is to
show that Option.fold(=>Z) preserves the by-nameness
of the zero value, which may or may not be in the
spec yet.  The addition to 6.26.5 was (I think)
rejected as implementation-specific; it should
use the wording from 6.6 about "preserving by-
nameness" or "by-namity" (the latter has the
advantage of sounding both more technical and
somewhat inane).